+++ TrelloDotNet/TrelloDotNet/Model/StickerDefaultImageId.cs
@@ -1,0 +1,61 @@
+﻿namespace TrelloDotNet.Model
+{
+    /// <summary>
+    /// Default Supported Stickers
+    /// </summary>
+    public enum StickerDefaultImageId
+    {
+        /// <summary>
+        /// Not a default Sticker
+        /// </summary>
+        NotADefault,
+        /// <summary>
+        /// check
+        /// </summary>
+        Check,
+        /// <summary>
+        /// heart
+        /// </summary>
+        Heart,
+        /// <summary>
+        /// warning
+        /// </summary>
+        Warning,
+        /// <summary>
+        /// clock
+        /// </summary>
+        Clock,
+        /// <summary>
+        /// smile
+        /// </summary>
+        Smile,
+        /// <summary>
+        /// laugh
+        /// </summary>
+        Laugh,
+        /// <summary>
+        /// hug
+        /// </summary>
+        Huh,
+        /// <summary>
+        /// frown
+        /// </summary>
+        Frown,
+        /// <summary>
+        /// thumbsup
+        /// </summary>
+        ThumbsUp,
+        /// <summary>
+        /// thumbsdown
+        /// </summary>
+        ThumbsDown,
+        /// <summary>
+        /// star
+        /// </summary>
+        Star,
+        /// <summary>
+        /// rocketship
+        /// </summary>
+        RocketShip
+    }
+}